HACK: Make Mali400 driver work on H3
It is not a very clean patch, but it works. There is no power
management and Mali is clocked at a relatively low speed (252 MHz)
in order to avoid excessive overheating. This is fast enough for
lima-memtester anyway.

The clock setup code is partially based on the Mali r4p0 driver
sources from "modules/mali/DX910-SW-99002-r4p0-00rel0"