WIP remove vsync_task from sunxi fb driver.
Gets rid of those uninterruptible "vsync proc" kernel threads, but is
surely only a workaround to hide the real problem.

This does not get the system load down to zero, as an
"usb-hardware-scan" kernel thread is still uninterruptible as well.

This is based on loboris' patch but tries to do it more cleanly.